feat: restructure Profile page into focused screens
- Break monolithic Profile page into four focused screens:
  - /profile: simplified user info (email, names, password)
  - /locations: location management with leave functionality
  - /checkout-history: comprehensive checkout/return history
  - /settings: dark/light mode toggle (moved from header)

- Update navigation dropdown with new menu items
- Remove dark mode toggle from header, consolidate in Settings
- Maintain all existing functionality with improved organization
- Implement consistent Material UI styling across new pages

Resolves GitHub issue #55
