+++ docs/reference/CHANGELOG.md
@@ -2,6 +2,34 @@
 
 This file documents all completed features, fixes, and improvements to the LibraryCard project.
 
+## July 14, 2025 - Profile Page Restructure Implementation
+
+### Complete Profile Page Restructure - GitHub Issue #55 RESOLVED!
+- **RESTRUCTURED**: Monolithic Profile page into four focused, single-purpose screens for improved UX and code organization
+- **CREATED**: Simplified Profile page (`/profile`) containing only essential user information: email, first name, last name, and password management
+- **EXTRACTED**: Locations page (`/locations`) with complete location management functionality including leave location capability and permission checks
+- **EXTRACTED**: Checkout History page (`/checkout-history`) displaying comprehensive checkout and return history with book details and action dates
+- **CREATED**: Settings page (`/settings`) with dark/light mode toggle functionality moved from header for better organization
+- **UPDATED**: Navigation dropdown in AppLayout to include new menu items (Locations, Checkout History, Settings) linked from user account menu
+- **REMOVED**: Dark mode toggle from header, consolidating theme controls into dedicated Settings page
+- **MAINTAINED**: All existing functionality while improving code organization and user experience through focused page responsibilities
+
+### Profile Page Restructure Technical Implementation
+- **SIMPLIFIED**: Profile page interface removing bio field implementation per user request to focus on core profile data
+- **MODULARIZED**: Profile functionality into dedicated pages with clear separation of concerns and improved maintainability
+- **ENHANCED**: Navigation structure with logical grouping of profile-related functions in dropdown menu
+- **PRESERVED**: All authentication flows, permission checks, and user data management while restructuring UI organization
+- **IMPLEMENTED**: Consistent Material UI styling and user experience patterns across all new profile-related pages
+- **OPTIMIZED**: User workflow by organizing related functionality into focused screens accessible through intuitive navigation
+
+### User Experience Improvements
+- **FOCUSED**: Profile page on essential user information (email, names, password) without distracting secondary features
+- **ORGANIZED**: Location management into dedicated page with proper permission handling and leave functionality
+- **CENTRALIZED**: Checkout history viewing in dedicated page with comprehensive activity tracking and book details
+- **CONSOLIDATED**: Application settings into focused Settings page with theme toggle and future setting expansion capability
+- **STREAMLINED**: Navigation flow with logical menu organization and consistent user experience across all profile-related screens
+- **ENHANCED**: Overall application organization by separating profile data management from location management and settings
+
 ## July 14, 2025 - Unnecessary Content Refresh Prevention System Implementation
 
 ### Complete Refresh Prevention System - GitHub Issue #50 RESOLVED!
